separated routes and controllers
